+++ backend/src/main/java/turip/auth/repository/RefreshTokenRepository.java
@@ -1,9 +1,9 @@
-package turip.member.repository;
+package turip.auth.repository;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
+import turip.auth.domain.RefreshToken;
 import turip.member.domain.Member;
-import turip.member.domain.RefreshToken;
 
 public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
