+++ backend/src/main/java/turip/auth/token/GoogleTokenParser.java
@@ -1,4 +1,4 @@
-package turip.auth;
+package turip.auth.token;
 
 import com.google.api.client.googleapis.auth.oauth2.GoogleIdToken;
 import com.google.api.client.googleapis.auth.oauth2.GoogleIdToken.Payload;
