+++ backend/src/main/java/turip/auth/token/JwtProvider.java
@@ -1,4 +1,4 @@
-package turip.auth;
+package turip.auth.token;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
