+++ backend/src/main/java/turip/common/configuration/WebMvcConfiguration.java
@@ -5,7 +5,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import turip.auth.AuthAccountArgumentResolver;
+import turip.auth.resolver.AuthAccountArgumentResolver;
 
 @Configuration
 @RequiredArgsConstructor
